==> Predicting the Bundesliga with ML/1_ETL_code.ipynb <==
### This notebook contains the Extraction, Transformation and Loading of the dataset steps.

#### Notebook 1: Extract-Transform-Load
#### Notebook 2: Data Visualization
#### Notebook 3: Feature Engineering, Hyperparameter tuning and Modelling
#### Notebook 4: Result Evaluation

import numpy as np
import pandas as pd

### Data on Bundesliga Game results was obtained from http://www.football-data.co.uk/germanym.php

##### The data was downloaded on my local machine and transformed from csv into regular excel files, here they are importet into python

file1 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0001.xlsx", header=True)
file1 = pd.read_excel(file1)
file2 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0102.xlsx", header=True)
file2 = pd.read_excel(file2)
file3 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0203.xlsx", header=True)
file3 = pd.read_excel(file3)
file4 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0304.xlsx", header=True)
file4 = pd.read_excel(file4)
file5 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0405.xlsx", header=True)
file5 = pd.read_excel(file5)
file6 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0506.xlsx", header=True)
file6 = pd.read_excel(file6)
file7 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0607.xlsx", header=True)
file7 = pd.read_excel(file7)
file8 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0708.xlsx", header=True)
file8 = pd.read_excel(file8)
file9 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0809.xlsx", header=True)
file9 = pd.read_excel(file9)
file10 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data0910.xlsx", header=True)
file10 = pd.read_excel(file10)
file11 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1011.xlsx", header=True)
file11 = pd.read_excel(file11)
file12 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1112.xlsx", header=True)
file12 = pd.read_excel(file12)
file13 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1213.xlsx", header=True)
file13 = pd.read_excel(file13)
file14 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1314.xlsx", header=True)
file14 = pd.read_excel(file14)
file15 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1415.xlsx", header=True)
file15 = pd.read_excel(file15)
file16 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1516.xlsx", header=True)
file16 = pd.read_excel(file16)
file17 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1617.xlsx", header=True)
file17 = pd.read_excel(file17)
file18 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1718.xlsx", header=True)
file18 = pd.read_excel(file18)
file19 = pd.ExcelFile(r"C:\Users\### LOCAL PATH ###\data1819.xlsx", header=True)
file19 = pd.read_excel(file19)

files = [file1, file2, file3, file4, file5, file6, file7, file8, file9, file10, file11, file12, file13, file14, file15, file16, file17, file18, file19]

# The data includes several acronyms
# Included are general statistics such as the goals scored by either team, team names, the result ('H' - Home team wins)
# Further included are the betting odds from by several sports betting providers
# Unfortunately, not every file includes every statistic
file18.tail(3)

##### Adding a column with the year that the season started, to which a game belongs

x = 2000

for file in files:
    
    file["Season start"] = x
    
    x +=1

##### The code below shortens the data aiming to keep as many datapoints as possible while keeping dimensionality high

file1 = file1[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file2 = file2[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file3 = file3[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file4 = file4[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file5 = file5[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file6 = file6[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR", "IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file7 = file7[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file8 = file8[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file9 = file9[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file10 = file10[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file11 = file11[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file12 = file12[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file13 = file13[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file14 = file14[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file15 = file15[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file16 = file16[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file17 = file17[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file18 = file18[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","LBH","LBD","LBA","WHH","WHD","WHA"]]
file19 = file19[["Season start","HomeTeam","AwayTeam","FTHG","FTAG","FTR","IWH","IWD","IWA","B365H","B365D","B365A","WHH","WHD","WHA"]]

files = [file1, file2, file3, file4, file5, file6, file7, file8, file9, file10, file11, file12, file13, file14, file15, file16, file17, file18]
file19 = file19.rename(columns={"B365H":"LBH","B365D":"LBD","B365A":"LBA"})# file19 slightly different since no odds available from a certain provider - replaced by another
files2 = [file18, file19]

## Feature Engineering part 1:
# - Adding the Gameday

for file in files:
    file["Gameday"] = np.nan
    x = 1
    for i in range(0,306,9):
        file.iloc[i:i+9,-1] = x
        x += 1

for file in files2:
    file["Gameday"] = np.nan
    x = 1
    for i in range(0,306,9):
        file.iloc[i:i+9,-1] = x
        x += 1
        

# combining the individual files

data = pd.DataFrame()

for file in files:
    data = data.append(file)
    
data_current_season = pd.DataFrame()
data_current_season = file18.append(file19)

##### Some data is still missing - mostly odds by one of the sports betting providers

data.shape

data.dropna().shape

data = data.reset_index(drop=True)
data_current_season = data_current_season.reset_index(drop=True)
data.head()

null_counts = data.isnull().sum()
null_counts[null_counts > 0].sort_values(ascending = False)

##### Averaging betting odds and finalizing the initial dataset

data["Odds H"] = data[["IWH","LBH","WHH"]].median(axis=1)
data["Odds D"] = data[["IWD","LBD","WHD"]].median(axis=1)
data["Odds A"] = data[["IWA","LBA","WHA"]].median(axis=1)

data_current_season["Odds H"] = data_current_season[["IWH","LBH","WHH"]].median(axis=1)
data_current_season["Odds D"] = data_current_season[["IWD","LBD","WHD"]].median(axis=1)
data_current_season["Odds A"] = data_current_season[["IWA","LBA","WHA"]].median(axis=1)

data = data.rename(columns={"FTHG":"Home Goals","FTAG":"Away Goals","FTR":"Result"})
data = data[["Season start","Gameday","HomeTeam","AwayTeam","Home Goals","Away Goals","Result","Odds H","Odds D","Odds A"]]

data_current_season = data_current_season.rename(columns={"FTHG":"Home Goals","FTAG":"Away Goals","FTR":"Result"})
data_current_season = data_current_season[["Season start","Gameday","HomeTeam","AwayTeam","Home Goals","Away Goals","Result","Odds H","Odds D","Odds A"]]


data.head()

data.describe()

# data that will be used for training and modeling
data.to_csv(r'C:\Users\### LOCAL PATH ###\dataforcapstone.txt', header = True, index=None, sep='\t')

# data for "reality check"
data_current_season.to_csv(r'C:\Users\### LOCAL PATH ###\data_current_season.txt', header = True, index=None, sep='\t')

data_current_season.head()
